chore: add CHANGELOG.md
The diff shows that a new file CHANGELOG.md was added to the codebase. Since this doesn't add new features or fix bugs, but rather adds a project file, the most appropriate type is 'chore'. The scope can be omitted since it's a new top-level file. The description concisely summarizes that the change was the addition of a CHANGELOG.md file.