feat(changelog): add changelog generation command
The code changes show:
- A new entry was added to the CHANGELOG.md file mentioning "add changelog generation command"
- In the gitwise/features/pr.py file, a new function `generate_pr_description` was added to generate a structured PR description from commits

Based on these changes, the most appropriate conventional commit message is:

feat(changelog): add changelog generation command

This succinctly describes the main change, which is adding a new feature for generating the changelog. The scope is "changelog" since that's the area of the codebase being modified.