"Merge pull request #44 from PayasPandey11/fix/ux"
This commit includes changes to the commit message format, which can be found in the [Git Commit Messages](https: //git-scm.com/docs/git-commit#Documentation/git-commit.txt--s) documentation. The commit message should be concise, informative, and describe the change being made clearly. The commit message should make reference to an issue or pull request that is included in the commit.