Here is a possible commit message based on the git diff you provided: `type: update`
`Description:`

We've updated `add.py` to display the current LLM backend to the user. Previously, this information was not displayed to the user. Additionally, we've added a new option to create a commit with these changes.

The reason for these changes is to provide more informative feedback to the user about the current LLM backend being used by Gitwise. By displaying this information, users can better understand their environment and make informed decisions about how to use Gitwise. The new option to create a commit with these changes allows users to easily document and save these changes for future reference.