chore: Add opt-out for PR creation after successful push
Previously, if the user opted not to create a PR after pushing, the
push feature would return False, indicating failure. However, opting
out of PR creation after a successful push should be considered a
successful operation.

Return True after a successful push if the user chooses not to
create a PR, properly reflecting the success status of the feature.

Fixes #217